requirements.txt
flask
flask_sqlalchemy
python-docx
sumy
scikit-learn
gunicorn
